Se solicita la extensión de audio y video. Añadida la posibilidad de no eliminar los archivos originales

+++ unir_video_y_audio.sh
@@ -21,23 +21,32 @@
 
 #!/bin/bash
 
+# Capturamos por teclado la extensión de video y audio a mezclar y la opción
+# de borrar los archvios originales.
+read -p "Introduce la extensión de los archivos de video a mezclar (sin punto)_ " video_ext
+read -p "Introduce la extensión de los archivos de audio a mezclar (sin punto)_ " audio_ext
+read -p "¿Deseas borrar los archivos originales? (s/n) " borrar_archivos
+
 echo "Recorremos el directorio actual buscando archivos de video..."
-for video in *.mp4
+for video in *.$video_ext
   do
-    echo "Pista de video -> ${video}"
-    # OBTENEMOS EL NOMBRE DEL FICHERO SIN EXTENSIÓN Y LA PISTA DE AUDIO
+    echo  "Pista de video -> ${video}"
+    # Del fichero de video eliminamos la extensión y construimos el nombre
+    # del fichero de audio
     filename=$(echo "$video" | cut -f 1 -d '.')
-    audio="$filename".aac
-    echo "Pista de audio -> ${audio}"
+    audio="$filename".$audio_ext
+    echo " Pista de audio -> ${audio}"
     if [ -f $audio ]; then
-      video_final="$filename"_final.mp4
-      # Pasamos los archivos de video y audio; de cada uno de ellos, nos aseguramos
+      video_final="$filename"_final.$video_ext
+      # Pasamos los archivos de video y audio y de cada uno de ellos, nos aseguramos
       # de obtener la pista correcta, porque un archivo puede contener varias
-      # pistas.
+      # pistas
       ffmpeg -i $video -i $audio -c copy -map 0:v:0 -map 1:a:0 $video_final
-      echo "Generado archivo ${video_final}"
-      # SE ELIMINAN LOS ARCHIVOS SEPARADOS
-      rm -f $video $audio
+      echo " Generado archivo ${video_final}"
+      # Se eliminan los archivos separados
+      if [[ ${borrar_archivos,,} == "s" ]]; then
+        rm -f $video $audio
+      fi
     else
       echo "No existe la pista de audio ${audio}"
     fi
